refactor(test): extract shared error-tracking fixtures to conftest.py
Extract `reset_error_tracking` and `temp_store` fixtures from
test_error_tracking.py and test_error_tracking_edge_cases.py into
a shared conftest.py to reduce duplication.

Changes:
- NEW: tests/vibe3/exceptions/conftest.py with shared fixtures
- REMOVE: duplicate fixture definitions from both test files
- CLEANUP: unused imports (Iterator, Path, pytest where applicable)

Closes #1439
